Cleaned code to abide by checkStyle rules
