+++ src/main/java/seedu/financeit/datatrackers/entrytracker/EntryTracker.java
@@ -8,9 +8,9 @@ import seedu.financeit.common.exceptions.InsufficientParamsException;
 import seedu.financeit.common.exceptions.ItemNotFoundException;
 import seedu.financeit.datatrackers.goaltracker.GoalTracker;
 import seedu.financeit.datatrackers.manualtracker.Ledger;
-import seedu.financeit.datatrackers.entrytracker.EntryCommands.CreateEntryCommand;
-import seedu.financeit.datatrackers.entrytracker.EntryCommands.EditEntryCommand;
-import seedu.financeit.datatrackers.entrytracker.EntryCommands.RetrieveEntryCommand;
+import seedu.financeit.datatrackers.entrytracker.commands.CreateEntryCommand;
+import seedu.financeit.datatrackers.entrytracker.commands.EditEntryCommand;
+import seedu.financeit.datatrackers.entrytracker.commands.RetrieveEntryCommand;
 import seedu.financeit.parser.InputParser;
 import seedu.financeit.ui.TablePrinter;
 import seedu.financeit.ui.UiManager;
