+++ src/main/java/seedu/financeit/datatrackers/manualtracker/ManualTracker.java
@@ -6,8 +6,8 @@ import seedu.financeit.common.exceptions.DuplicateInputException;
 import seedu.financeit.common.exceptions.InsufficientParamsException;
 import seedu.financeit.common.exceptions.ItemNotFoundException;
 import seedu.financeit.datatrackers.goaltracker.GoalTracker;
-import seedu.financeit.datatrackers.manualtracker.ManualTrackerCommands.CreateLedgerCommand;
-import seedu.financeit.datatrackers.manualtracker.ManualTrackerCommands.RetrieveLedgerCommand;
+import seedu.financeit.datatrackers.manualtracker.commands.CreateLedgerCommand;
+import seedu.financeit.datatrackers.manualtracker.commands.RetrieveLedgerCommand;
 import seedu.financeit.datatrackers.entrytracker.EntryTracker;
 import seedu.financeit.parser.InputParser;
 import seedu.financeit.ui.TablePrinter;
